Adding component Load functionality
To add this functionality adapted to the new systemc approach
I had to adapt the other components.
HardwareComponentXmlParser - is now able to load components
and convert them to a set of HardwareComponentInfo that will
be used by the ComponentDatabase.
HardwareReconfiguration - Was adapted to support the new signature
methods from HardwareComponentXmlParser.
HardwareComponent - Added new fields to the HardwareComponentInfo
structure.