Checks to recognize already bound interfaces
- HardwareComponent
The interface binding using only the ports was not working so I decided
to use signals to it. Thus the get_interface method will work even
without the need of elaborate the components.

When using this approach was needed add checks to avoid using new
instance of signals in the same component, what causes an exception in
systemC.